refactor code
refactor code that creates the file to upload to s3
